Move the playground instance indicators to the Blueprint tab button
