+++ client/Sidebar/BlueprintPanel.jsx
@@ -1,5 +1,4 @@
 import { useState, useMemo } from 'react';
-import { createPortal } from 'react-dom';
 import { toast } from 'react-toastify';
 import { useAppState } from '../context/AppStateContext.jsx';
 import { blueprintToForm, useBlueprintFormState } from '../state/useBlueprintFormState.js';
@@ -11,49 +10,12 @@ import { SlugListSection } from '../blueprint/SlugListSection.jsx';
 import { ContentSection } from '../blueprint/ContentSection.jsx';
 import { SectionBadge } from './SectionBadge.jsx';
 
-const SLOT_STATUS_LABELS = {
-  active: 'In use',
-  booting: 'Loading',
-  idle: 'Idle',
-  warm: 'Ready and warm',
-};
-
-function statusLabel(status) {
-  return SLOT_STATUS_LABELS[status] ?? 'Unknown';
-}
-
-function poolSlotTitle(slot) {
-  const port = slot.port ? ` (port ${slot.port})` : '';
-  return `Playground ${slot.index + 1}${port}: ${statusLabel(slot.status)}`;
-}
-
-function poolSlotClass(status) {
-  return SLOT_STATUS_LABELS[status] ? status : 'unknown';
-}
-
-function inferredPoolSlots(poolStatus) {
-  if (Array.isArray(poolStatus.slots) && poolStatus.slots.length > 0) {
-    return poolStatus.slots;
-  }
-
-  const total = Number(poolStatus.total) || 0;
-  const warm = Number(poolStatus.warm) || 0;
-  const booting = Number(poolStatus.booting) || 0;
-
-  return Array.from({ length: total }, (_, index) => ({
-    index,
-    port: null,
-    status: index < warm ? 'warm' : index < warm + booting ? 'booting' : 'idle',
-  }));
-}
-
 export function BlueprintPanel({ embedded = false } = {}) {
-  const { blueprint: appBlueprint, defaultBlueprint, setBlueprint, isBlueprintModified, poolStatus } = useAppState();
+  const { blueprint: appBlueprint, defaultBlueprint, setBlueprint, isBlueprintModified } = useAppState();
   const { formState, updateForm, loadBlueprint, compiledBlueprint } =
     useBlueprintFormState(appBlueprint);
 
   const [isSaving, setIsSaving] = useState(false);
-  const [poolTooltip, setPoolTooltip] = useState(null);
 
   // Compare on form-state (not compiled JSON) so incidental key-order / shape
   // differences in the on-disk blueprint don't make the form look "dirty".
@@ -67,7 +29,6 @@ export function BlueprintPanel({ embedded = false } = {}) {
     () => JSON.stringify(formState) === JSON.stringify(defaultForm),
     [formState, defaultForm],
   );
-  const poolSlots = useMemo(() => inferredPoolSlots(poolStatus), [poolStatus]);
 
   async function handleSave() {
     setIsSaving(true);
@@ -95,22 +56,6 @@ export function BlueprintPanel({ embedded = false } = {}) {
     } finally {
       setIsSaving(false);
     }
-  }
-
-  function showPoolTooltip(text, target) {
-    const rect = target.getBoundingClientRect();
-    const minX = 140;
-    const maxX = Math.max(minX, window.innerWidth - minX);
-    const centerX = rect.left + rect.width / 2;
-    setPoolTooltip({
-      text,
-      left: Math.min(Math.max(centerX, minX), maxX),
-      top: rect.top - 8,
-    });
-  }
-
-  function hidePoolTooltip() {
-    setPoolTooltip(null);
   }
 
   const content = (
@@ -141,26 +86,6 @@ export function BlueprintPanel({ embedded = false } = {}) {
       </section>
 
       <div className="blueprint-panel-actions">
-        {poolSlots.length > 0 && (
-          <div className="blueprint-pool-indicators" aria-label="Playground pool status">
-            {poolSlots.map((slot) => {
-              const title = poolSlotTitle(slot);
-              return (
-                <span
-                  key={slot.port ?? slot.index}
-                  className={`blueprint-pool-indicator blueprint-pool-indicator--${poolSlotClass(slot.status)}`}
-                  aria-label={title}
-                  role="img"
-                  tabIndex={0}
-                  onBlur={hidePoolTooltip}
-                  onFocus={(event) => showPoolTooltip(title, event.currentTarget)}
-                  onMouseEnter={(event) => showPoolTooltip(title, event.currentTarget)}
-                  onMouseLeave={hidePoolTooltip}
-                />
-              );
-            })}
-          </div>
-        )}
         <button
           type="button"
           className="bp-action-btn bp-action-btn--ghost"
@@ -178,19 +103,6 @@ export function BlueprintPanel({ embedded = false } = {}) {
           {isSaving ? 'Saving…' : 'Save'}
         </button>
       </div>
-      {poolTooltip && createPortal(
-        <div
-          className="blueprint-pool-tooltip"
-          role="tooltip"
-          style={{
-            left: `${poolTooltip.left}px`,
-            top: `${poolTooltip.top}px`,
-          }}
-        >
-          {poolTooltip.text}
-        </div>,
-        document.body,
-      )}
     </div>
   );
